% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipartplot.R
\name{bipartplot}
\alias{bipartplot}
\title{bipartplot}
\usage{
bipartplot(edgelist, return_ggplot = FALSE)
}
\arguments{
\item{edgelist}{The edgelist corresponding to the bipartite network}

\item{return_ggplot}{Flag to return a ggplot2 object instead of plotting}
}
\description{
Takes an edge list (from a bipartite network)
And plots a railway bipartite plot
using ggplot2
}
