% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{cellphone_db_links}
\alias{cellphone_db_links}
\title{Ligand-receptor interaction dataframe}
\format{data frame}
\usage{
data(ramilowski_links)
}
\description{
Dataframe of ligand-receptor interactions curated by Teichmann et al
https://www.cellphonedb.org/
}
\keyword{datasets}
