% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{degs}
\alias{degs}
\title{Sample genes for example}
\format{data frame}
\usage{
data(degs)
}
\description{
Sets of differentially expressed genes (DEGs) from two
different studies: GSE110226 (Alzheimer's disease DEGs in the brain)
and GSE61490(Chronic Periodontitis DEGs in monocytes)
}
\keyword{degs}
