% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kxk_test.R
\name{kxk_test}
\alias{kxk_test}
\title{kxk_test}
\usage{
kxk_test(n)
}
\arguments{
\item{n}{The number of links found}
}
\value{
Prints and estimate of the probability o
}
\description{
Uses a 100 000 simulation of randomly sampled genes from a 13000 gene pool to
obtain estimates of links by random selection of genes.
This estimates are valid only for 1000 receptors and 1000 ligands.
}
