% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ramilowski_links}
\alias{ramilowski_links}
\title{Ligand-receptor interaction dataframe}
\format{data frame}
\usage{
data(ramilowski_links)
}
\description{
Dataframe of ligand-receptor interactions curated by Ramilowski et al, 2015
Ref: https://www.nature.com/articles/ncomms8866
}
\references{
Ramilowski, Jordan A., et al. "A draft network of
ligand–receptor-mediated multicellular signalling in human."
Nature communications 6 (2015): 7866.
\href{https://www.ncbi.nlm.nih.gov/pubmed/26198319}{PubMed}
}
\keyword{datasets,}
\keyword{links,}
\keyword{ramilowski}
