% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zelda.R
\name{zelda}
\alias{zelda}
\title{zelda}
\usage{
zelda(receptors, ligands, database = "ramilowski_links",
  return_ggplot = FALSE, plot_it = TRUE, return_links = FALSE)
}
\arguments{
\item{receptors}{A list of genes containing putative receptors}

\item{ligands}{A list of genes containg putative ligands}

\item{database}{The database to be used to find the links. Defaults to "ramilowski_links".
Other options: "cellphone_db_links"}

\item{return_ggplot}{Flag to return a ggplot2 object instead of plotting. Defaults to FALSE.}

\item{plot_it}{Flag to plot bipartite network or not. Defaults to TRUE.}

\item{return_links}{Flag to return a dataframe with the links object instead of plotting. Defaults to FALSE.}
}
\description{
Takes two lists of genes and finds the receptor-ligand relations (links) between them.
Plots a bipartite plot of the found genes.
Returns a bipartite network of the connections.
}
\examples{
data(degs)
degs_alzheimer <- degs$degs_alzheimer
degs_periodontitis <- degs$degs_periodontitis
p <- zelda(receptors = degs_alzheimer, ligands = degs_periodontitis, return_ggplot = TRUE, plot_it = FALSE)
p
links <- p <- zelda(receptors = degs_alzheimer, ligands = degs_periodontitis, plot_it = FALSE, return_links = TRUE)
head(links)
}
