% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zelda.R
\name{plot_links}
\alias{plot_links}
\title{plot_links}
\usage{
plot_links(edgelist, return_ggplot = FALSE)
}
\arguments{
\item{edgelist}{The edgelist corresponding to the bipartite network}

\item{return_ggplot}{Flag to return a ggplot2 object instead of plotting}
}
\description{
Takes an edge list (from a bipartite network)
And plots a railway bipartite plot using ggplot2
}
\details{
The link list can be regarded as a bipartite network.
Only the first two columns of the edgelist are used
}
\examples{
 receptor_candidates <- c("CD3", "AIF", "CCR3", "CXCR3")
 ligand_candidates <- c("TP53", "NFKB1", "CXCL9")
 links <-
   return_links(receptors = receptor_candidates,
                ligands = ligand_candidates,
                database = "all")
 head(links)

 plot_links(links)
}
