% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zelda.R
\name{return_links}
\alias{return_links}
\title{return_links}
\usage{
return_links(receptors, ligands, database = "ramilowski_links")
}
\arguments{
\item{receptors}{A list of genes containing putative receptors}

\item{ligands}{A list of genes containg putative ligands}

\item{database}{The database to be used to find the links. Defaults to "ramilowski_links".
Other options: "cellphone_db_links"; "all"}
}
\value{
Returns a dataframe with the lists found in the databases.
}
\description{
zelda Takes two lists of genes and finds the receptor-ligand relations (links) between them.
Plots a bipartite plot of the found genes.
Returns a bipartite network of the connections.
}
\examples{

 receptor_candidates <- c("CD3", "AIF", "CCR3", "CXCR3")
 ligand_candidates <- c("TP53", "NFKB1", "CXCL9")
 links <- return_links(receptor_candidates, ligand_candidates, database = "all")
 head(links)
}
